requirements.txt
django==3.2.5
Pillow==8.3.1
django-braces==1.14.0
django-embed-video==1.4.0
python-memcached==1.59.0
django-memcache-status==2.2.0
djangorestframework==3.12.4